Update us_plot to add zoom, pan, print, and svg functions Add us_buffer Numerous other smaller corrections

+++ trunk/gui/us_selectmodel10.h
@@ -16,10 +16,10 @@ class US_EXTERN US_SelectModel10 : public US_WidgetsDialog
       //! \param temp_stoich1 - Location for the stoichiometry value
       //! \param  p           - The parent widget
       //! \param  f           - Window flags to be passed to the widget  
-      US_SelectModel10( float&, QWidget* = 0, Qt::WindowFlags = 0 );
+      US_SelectModel10( double&, QWidget* = 0, Qt::WindowFlags = 0 );
 
    private:
-      float&     stoich1;
+      double&    stoich1;
       QLineEdit* le_stoich1;
       US_Help    showhelp;
